Fix bug in creating `global` list and make `buildStructureEnsembles` more robust
`buildStructureEnsembles` was relying on a certain number of levels in the file path name. It wasn't robust to if a user had a different number of parent directories in `dataFolder`. Now it should be-- to find the ensemble name, it looks for the directory in the path named "historical" or "rcp85" and then looks two directories below that for the name