Close #4
I subsetted out only the following parts of the original sample files:

- cities: Only "balt", "ny", "nwk", "phil", "prov"
- grid points (latlong): Only ones closest to these cities (400, 401, 402)
- times: historical: only since 1980 (no change for rcp8.5)
- tas: Only the subset of tas that met these conditions (grid points 400, 401, 402, historical since 1980, anytime for rcp8.5)

I saved these in the `inst` directory so we could save them as .csv files, so they will work appropriately with the functions. If we put them in a `data` directory, we need to save them as .Rdata files and the functions would not work correctly.