Make `gen_hw_set` return model info
Colin: This is just a thought-- in addition to writing out the model info, we could have `gen_hw_set`'s return value be the dataframe of model information. That would be a more interesting return value than a 0 / 1 flag, since it's likely the first thing that a researcher would want to check to make sure everything worked like he wanted.