Work on `checkCustomBounds`.
Note to Colin: I change the parameter value in this function from `length` to `expected_length` to prevent any confusion between the argument value and the base R function with the same name. In R style, it's considered good practice to avoid naming any objects (including as arguments in function calls) with the same name as existing functions.