Create new function to check custom date boundaries.
If I can get this to work correctly, this will replace the current `checkCustomBounds`. It takes a vector of length 2, with the first value the lower bound and the second an upper bound on the date. The restrictions are: (1) vector class must be numeric; (2) bounds cannot span 2005 (so they must be a period either completely in "historic" or completely in "rcp85"; and (3) they cannot go before 1980 or after 2099, so we can ensure we'll have data for the period requested.