Close #78
The old version was looking for the index that was 1 minus the end year + 1 (so, the last date before the first that matched to the next year). This failed if you were trying to go right up to the last year of the dataset. The new version can handle that case.