Change `as.numeric` in IDHeatwavesCPPwrapper
Previously, this was `as.integer`. However, occasionally on Windows and Linux R builds, tests with this function cause a crash, with the error:

Error in IDHeatwavesCPP(heatwaveLength = as.integer(numDays), tempsExceedingCutoff = tempsExceedingthreshold) :
     unimplemented type 'NULL' in 'coerceToReal'

This bug does not always reproduce. There is some more information on this issue when using Rcpp here:

https://github.com/RGLab/COMPASS/issues/8