Close #144 by using `IntegerVector` instead of `Numeric`
Before, we were wrapping `hw` and `hwNumber`, which are integer vectors, and passing to `NumericVector`s. This usually did not cause a bug, but every now and then would on Windows and Linux and caused CRAN errors once (> 1 month after posting on CRAN). The error message came from  a call to `coerceToReal`, which I think might have happened in coercing these integer vectors to numerics to put them in `NumericVectors`. All tests and checks are running correctly now.