Fix Feb. 30 bug.
Evidently, some climate models include Feb. 30 (?!) and Feb. 29 on non-leap years. This was mucking up some of the date conversions in the code and causing errors with a couple of the CMIP5 models that do this. The package now can run with heatwaves that start on Feb. 29 in a non-leap year or Feb. 30. In these cases, the starting day of the year is re-assigned to Mar.1 of the same year, and the starting month is assigned as March.