Remove `input_metric` parameter
This helps make the package more generalizable to things other than temperature--- the user now will always get output in the same metric used for the input (e.g., if the input files are in Kelvin, any absolute measures in the output files will also be in Kelvin). Also, now all optional parameters (e.g., the thresholds for absolute metrics) should all be in the metric used in input files.